Rename for App Screenshare; use agora sample audio file; audio mixing sample refactor
