+++ API-Example-Unity/Assets/API-Example/app-screenshare-sample.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ac2f6b142e9c41cd9c2f9b5ed6f96a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
