+++ API-Example-Unity/Assets/API-Example/app-screenshare-sample/AgoraScreenShare.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49077210539eb471c84c3f0620d78fc8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
