+++ API-Example-Unity/Assets/API-Example/audio-mixing-sample/AudioMixing.cs
@@ -22,6 +22,7 @@ public class AudioMixing : MonoBehaviour
     private Logger logger;
     private IRtcEngine mRtcEngine = null;
     private IAudioPlaybackDeviceManager manager = null;
+    const string SampleAudioSubpath = "/audio/Agora.io-Interactions.mp3";
 
     // Start is called before the first frame update
     void Start()
@@ -70,9 +71,9 @@ public class AudioMixing : MonoBehaviour
 
 #if UNITY_ANDROID && !UNITY_EDITOR
         // On Android, the StreamingAssetPath is just accessed by /assets instead of Application.streamingAssetPath
-        localPath = "/assets/audio/DESERTMUSIC.wav";
+        localPath = "/assets" + SampleAudioSubpath;
 #else
-        localPath = Application.streamingAssetsPath + "/audio/" + "DESERTMUSIC.wav";
+        localPath = Application.streamingAssetsPath + SampleAudioSubpath;
 #endif
         logger.UpdateLog(string.Format("the audio file path: {0}", localPath));
 
@@ -95,7 +96,11 @@ public class AudioMixing : MonoBehaviour
     {
         Debug.Log("Playing with " + ( _useURL? "URL" : "local file") );
 
-        mRtcEngine.StartAudioMixing( _useURL? Sound_URL : localPath, true, false, -1, 0);
+        mRtcEngine.StartAudioMixing( filePath: _useURL? Sound_URL : localPath, 
+	                                 loopback: false, 
+				                      replace: true, 
+				                        cycle: -1, 
+					                 startPos: 0);
     }
 
     void StartAudioPlaybackTest()
