+++ API-Example-Unity/Assets/API-Example/audio-mixing-sample/AudioMixingScene.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8cb5873d5ce2441c58f975843144f122
+guid: c48b5ff915b3f43248c9b73de577f046
 DefaultImporter:
   externalObjects: {}
   userData:
