+++ API-Example-Unity/Assets/API-Example/custom-audio-source/Agora.io-Interactions.mp3.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5121197de53614c33810999172289239
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
