--- API-Example-Unity/Assets/API-Example/screenshare-sample/MainScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c57883c4b54ea4984b722542d3f4a76f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
