+++ API-Example-Unity/Assets/StreamingAssets/audio/Agora.io-Interactions.mp3.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 980b1cf6b7720460b85adf2d7f8697c7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
