--- API-Example-Unity/Assets/StreamingAssets/audio/DESERTMUSIC.wav.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 501687ec90aee4a50abb5fcde23b3704
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
